// src/pages/OrderModel/OrderModel.interface.ts
export enum sortTitles {
	CHECK = 'CHECK',
	ID = 'ID',
	COMMENT = 'COMMENT',
	STATUS = 'STATUS',
	PRICE = 'PRICE',
	NAME = 'NAME',
	SERVICE_TITLE = 'SERVICE_TITLE',
	DATE_UPDATED = 'DATE_UPDATED',
	DATE_CREATED = 'DATE_CREATED',
}

export enum sortTitlesView {
	CHECK = ' ',
	ID = '№',
	COMMENT = 'Комментарий',
	STATUS = 'Статус',
	PRICE = 'Цена',
	NAME = 'Имя заказчика',
	SERVICE_TITLE = 'Название услуги',
	DATE_UPDATED = 'Дата обновления',
	DATE_CREATED = 'Дата создания',
}

export enum sortTitlesSend {
	PRICE = 'price',
}